- 명령어 분리 - 패키지 정보 업데이트 - 가변 처리가 가능하도록 변경
