+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ms2-discord-bot",
   "version": "1.0.0",
-  "main": "src/index.mjs",
+  "main": "src/app.mjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -22,5 +22,13 @@
     "babel-plugin-wildcard": "^7.0.0",
     "eslint": "^8.8.0"
   },
-  "description": ""
+  "description": "메이플스토리 2의 OX 퀴즈쇼의 족보 정보를 제공하는 Discord 봇의 JavaScript 버전",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/shin6949/ms2-ox-bot-js.git"
+  },
+  "bugs": {
+    "url": "https://github.com/shin6949/ms2-ox-bot-js/issues"
+  },
+  "homepage": "https://github.com/shin6949/ms2-ox-bot-js#readme"
 }
