+++ src/app.mjs
@@ -1,23 +1,16 @@
 import dotenv from 'dotenv'
 import {Client, Collection, Intents} from "discord.js";
-import { default as command} from './commands/report-command.mjs';
-// import * as commands from 'commands';
-// 상수들을 가져옴
-import fs from 'fs';
+import * as commands from './commands/index.mjs';
 
 // Dev, Prod 구분 코드 필요.
 dotenv.config({ path: '../.env'});
 const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
 
+// commands의 파일을 불러와 명령어를 등록
 client.commands = new Collection();
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.mjs'));
-
-client.commands.set(command.data.name, command);
-// for (const file of commandFiles) {
-// 	const command = require(`./commands/${file}`);
-// 	console.log(command);
-// 	client.commands.set(command.data.name, command);
-// }
+for (const commandName of Object.keys(commands)) {
+	client.commands.set(commands[commandName].data.name, commands[commandName]);
+}
 
 client.once('ready', () => {
 	console.log('Ready!');
@@ -28,6 +21,7 @@ client.on('interactionCreate', async interaction => {
 	const command = client.commands.get(interaction.commandName);
 	if (!command) return;
 
+	// 불러온 명령어를 기반으로
 	try {
 		await command.execute(interaction);
 	} catch (error) {
@@ -35,4 +29,5 @@ client.on('interactionCreate', async interaction => {
 		await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
 	}
 });
-client.login(process.env.DISCORD_BOT_TOKEN);
+
+client.login(process.env.DISCORD_BOT_TOKEN).then(r => console.error(r));
