--- src/test.mjs
@@ -1,5 +1,0 @@
-// import { default as command} from './commands/report-command.mjs';
-import {command} from 'commands';
-
-console.log(typeof(command));
-console.log(command);
